continue with rank set/list/group experimentation
