+++ src/control/lib/hostlist/hostgroup.go
@@ -1,5 +1,5 @@
 //
-// (C) Copyright 2019 Intel Corporation.
+// (C) Copyright 2019-2020 Intel Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -28,6 +28,7 @@ import (
 	"sort"
 )
 
+// HostGroups maps a set of hosts to a string key value.
 type HostGroups map[string]*HostSet
 
 func (hg HostGroups) Keys() []string {
